+++ src/Enums/NetworkExchange/DocumentTypeCodes.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Enlivy\Enums\NetworkExchange;
+
+use Enlivy\Enums\Concern\EnumValues;
+
+enum DocumentTypeCodes: string
+{
+    use EnumValues;
+
+    case COMMERCIAL_INVOICE = '380';
+    case CREDIT_NOTE = '381';
+}